#80 [REFACTOR] 비즈니스 로직을 포함한 Result 클래스들을 도메인 패키지로 이전
