--- src/main/java/com/wakeUpTogetUp/togetUp/api/mission/model/CustomAnalysisResult.java
@@ -1,10 +1,0 @@
-package com.wakeUpTogetUp.togetUp.api.mission.model;
-
-public abstract class CustomAnalysisResult {
-
-    private String name;
-
-    private double confidence;
-
-    private BoundingBox box;
-}
